+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 **Unreleased**
 --------------
 
+0.2.1
+-----
+
+_2024-08-08_
+
 - **Fix:** Don't cache symbols between processing rounds. This better supports KSP2.
 - **Fix:** Workaround Kotlin plugin option parsing limitations. Contributing annotations should now be colon-delimited, and the underlying KSP argument is changed to a more consistent `anvil-ksp-extraContributingAnnotations`.